+++ ERPAPI/Models/Tax.cs
@@ -21,7 +21,7 @@ namespace ERPAPI.Models
         public string Description { get; set; }
         public Int64 IdEstado { get; set; }
         public string Estado { get; set; }
-        public double TaxPercentage { get; set; }
+        public decimal TaxPercentage { get; set; }
         public DateTime FechaCreacion { get; set; }
         public DateTime FechaModificacion { get; set; }
         public string UsuarioCreacion { get; set; }
